+++ src/components/content/care-team/patient-careteam-filters.tsx
@@ -1,29 +1,35 @@
 import { CareTeamModel } from "@/fhir/models/careteam";
+import { CareTeamPractitionerModel } from "@/fhir/models/careteam-practitioner";
 import { ResourceMap } from "@/fhir/types";
-import { isEqual, uniqWith } from "@/utils/nodash";
+import { orderBy, uniqBy } from "@/utils/nodash";
 
 export const applyCareTeamFilters = (
   data: fhir4.CareTeam[],
   includedResources: ResourceMap
 ) => {
-  const careTeamModel = data.map(
+  const careTeamModels = data.map(
     (careteam) => new CareTeamModel(careteam, includedResources)
   );
-  // Only want to obtain records that have a period.start but no period.end
-  const currentCareTeam = careTeamModel.filter(
-    (careTeam) => !careTeam.periodEnd
-  );
-  const careTeamData = uniqWith(currentCareTeam, (a, b) =>
-    isEqual(valuesToDedupeOn(a), valuesToDedupeOn(b))
+
+  const careTeamPractitionerModels: CareTeamPractitionerModel[] = [];
+  careTeamModels.forEach((careTeam) =>
+    careTeam.resource.participant?.forEach((participant) => {
+      const practitioner = careTeam.getPractitionerByID(
+        participant.member?.reference as string
+      );
+      if (practitioner?.id && participant.member?.reference) {
+        careTeamPractitionerModels.push(
+          new CareTeamPractitionerModel(careTeam, practitioner)
+        );
+      }
+    })
   );
 
-  return careTeamData;
+  const sortedCareTeamPractitionerModels = orderBy(
+    careTeamPractitionerModels,
+    ["effectiveStartDate", "id"],
+    ["desc", "desc"]
+  );
+
+  return uniqBy(sortedCareTeamPractitionerModels, "id");
 };
-
-const valuesToDedupeOn = (careteam: CareTeamModel) => [
-  careteam.status,
-  careteam.periodStart,
-  careteam.role,
-  careteam.practitionerQualification,
-  careteam.includedPerformer,
-];