+++ src/components/content/conditions/helpers/columns.tsx
@@ -21,9 +21,7 @@ export const patientConditionsColumns: TableColumn<ConditionModel>[] = [
     minWidth: 320,
     render: (condition) => (
       <div>
-        <div className="ctw-pc-title group-hover:ctw-underline">
-          {condition.display}
-        </div>
+        <div className="group-hover:ctw-underline">{condition.display}</div>
         <div className="ctw-pc-chapter">{condition.ccsChapter}</div>
       </div>
     ),
@@ -40,8 +38,8 @@ export const patientConditionsColumns: TableColumn<ConditionModel>[] = [
         >
           &bull;
         </div>
+        <div>{condition.displayStatus}</div>
         <div className="ctw-pc-status-and-extra">
-          <div>{condition.displayStatus}</div>
           <div>
             Last Updated:{" "}
             {compact([