+++ src/components/core/filter-bar/filter-bar-pills/select-pill.tsx
@@ -1,10 +1,8 @@
+import { faTrash, IconDefinition } from "@fortawesome/free-solid-svg-icons";
 import cx from "classnames";
 import { ReactNode } from "react";
-import {
-  FilterItemStatus,
-  FilterOptionSelect,
-} from "@/components/core/filter-bar/filter-bar-types";
-import { getIcon } from "@/components/core/filter-bar/filter-bar-utils";
+import { MenuItem } from "../../menu/menu-item";
+import { FilterOptionSelect } from "@/components/core/filter-bar/filter-bar-types";
 import {
   ListBox,
   ListBoxOptionStatus,
@@ -19,23 +17,20 @@ type FilterBarSelectPillProps = {
 };
 
 const buttonClassName =
-  "ctw-flex ctw-items-center ctw-capitalize ctw-text-content-black ctw-bg-bg-dark ctw-text-sm ctw-rounded ctw-my-2 ctw-py-2 ctw-px-3 ctw-relative ctw-mr-1 ctw-cursor-pointer ctw-border-0 ctw-border-transparent";
+  "ctw-flex ctw-items-center ctw-capitalize ctw-text-content-black ctw-bg-bg-dark ctw-text-sm ctw-rounded ctw-py-2 ctw-px-3 ctw-relative ctw-cursor-pointer ctw-border-0 ctw-border-transparent";
 const renderDisplay = (
-  display: string | ((status: FilterItemStatus) => string | ReactNode)
+  display: string | ((status: ListBoxOptionStatus) => string | ReactNode)
 ) => (isFunction(display) ? display({ active: false }) : display);
-const builtInButton = (name: string, key: string, icon: string) => ({
+const builtInButton = (name: string, key: string, icon: IconDefinition) => ({
   key,
   display: ({ listView, filter }: ListBoxOptionStatus) =>
     listView ? (
-      <span>
-        {getIcon(icon)} {name}
-      </span>
+      <MenuItem icon={icon}>{name}</MenuItem>
     ) : (
       filter && renderDisplay(filter.display)
     ),
 });
-const resetButton = builtInButton("reset filter", "_reset", "reset");
-const removeButton = builtInButton("remove filter", "_remove", "trash");
+const removeButton = builtInButton("remove filter", "_remove", faTrash);
 
 export function FilterBarSelectPill({
   filter,
@@ -45,31 +40,31 @@ export function FilterBarSelectPill({
   const filterNames = filter.values.map((value) =>
     isString(value) ? value : value.display
   );
-  const items = [...filterNames, resetButton, removeButton];
+  const items = [...filterNames, removeButton];
   return (
     <ListBox
-      useBasicStyles
       defaultIndex={items.length - 1} // clear index
       btnClassName={buttonClassName}
       optionsClassName="ctw-capitalize"
       items={items.map((item) => ({
         // eslint-disable-next-line react/no-unstable-nested-components
         display: ({ listView }: ListBoxOptionStatus) => {
-          if (item === resetButton || item === removeButton) {
+          if (item === removeButton) {
             return item.display({ listView, filter });
           }
           return listView ? (
             <span className="ctw-whitespace-nowrap">{item}</span>
           ) : (
-            <>
-              {renderDisplay(filter.display)}: {item}
-            </>
+            <div className="ctw-truncate ctw-font-medium ctw-text-content-black">
+              {renderDisplay(filter.display)}
+              <span className="ctw-font-normal">: {item}</span>
+            </div>
           );
         },
         key: isString(item) ? item : item.key,
         className: cx("ctw-capitalize", {
-          "ctw-border ctw-capitalize ctw-border-solid ctw-border-divider-light ctw-border-b-0":
-            item === resetButton,
+          "ctw-border ctw-capitalize ctw-border-solid ctw-border-divider-light ctw-border-0 ctw-border-t":
+            item === removeButton,
         }),
       }))}
       onChange={(index, item) => {